v0.1.2
Interaction with SQL database successfully added. Users can create new recipes, look at a full list of recipes in the database, or search by name, tag, or servings for specific recipes that match search criteria. Search bar in the header also functions as a search for a specific recipe name.