v0.1.3
Added ingredients and directions as columns in the database.
Added view.php which allows you to view all of the information regarding a recipe in a universal, appealing format by querying the database.
Added the ability to remove a recipe from the database by accessing its page.

Next update will look into error handling for form input. Currently urlencode and urldecode are effectively handling url-based error as the encoded version of the recipe name and tags are being stored in the database but there are additional measures that can be made for security and stability.

Additionally I want to look into restricting submissions with duplicate names, providing clear and descriptive error messages to the user, and better formatting view.php.

I also think that more can be done visually with view.php, I would love to give the user the ability to upload an image with their recipe to be displayed.

I'm also looking into exporting/importing recipes from csv so that I will have a larger dataset to work with.