v0.1.4
Polished display of ingredients and directions. Ingredients are displayed as an unordered list and directions are displayed as an ordered list. On the form, new list elements are distinguished using the newline character. Additionally, there is now a file input element that is used to display an image with each recipe on view.php

Additional error checking for the images/file input will be added in future versions(checking mime type, file size, removal of temporary files when the entry is removed from the database, etc.)